requirements.txt
asgiref==3.3.1
backcall==0.2.0
boto3==1.16.57
botocore==1.19.57
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
Django==3.1.5
django-storages==1.11.1
djangorestframework==3.12.2
idna==2.10
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.0
jupyter-client==6.1.11
jupyter-core==4.7.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
nba-api==1.1.9
numpy==1.19.5
pandas==1.2.0
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.10
psycopg2==2.8.6
Pygments==2.7.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
pywin32==300
pyzmq==21.0.1
requests==2.25.1
s3transfer==0.3.4
scikit-learn==0.24.0
scipy==1.6.0
six==1.15.0
SQLAlchemy==1.3.22
sqlparse==0.4.1
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
